requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.9.0
jinja2==3.1.4
feedparser==6.0.11
newspaper3k==0.2.8
schedule==1.2.0
python-dotenv==1.0.1
deepl==1.16.1